+++ system-tests/cucumber-maven-tia/maven-project/src/main/java/hellocucumber/Calculator.java
@@ -1,0 +1,11 @@
+package hellocucumber;
+
+public class Calculator {
+	public int add(int a, int b) {
+		return a + b;
+	}
+
+	public int subtract(int a, int b) {
+		return a - b;
+	}
+}